refactor(nx2-form): extract the form editor into editor.js
